New simpler bugfix for UPDATE and virtual BLOBs
When updating a table with virtual BLOB columns, the following might happen:
- an old record is read from the table, it has no virtual blob values
- update_virtual_fields() is run, vcol blob gets its value into the
  record. But only a pointer to the value is in the table->record[0],
  the value is in Field_blob::value String (but it doesn't have to be!
  it can be in the record, if the column is just a copy of another
  columns: ... b VARCHAR, c BLOB AS (b) ...)
- store_record(table,record[1]), old record now is in record[1]
- fill_record() prepares new values in record[0], vcol blob is updated,
  new value replaces the old one in the Field_blob::value
- now both record[1] and record[0] have a pointer that points to the
  *new* vcol blob value. Or record[1] has a pointer to nowhere if
   Field_blob::value had to realloc.

To fix this I have introduced a new String object 'read_value' in Field_blob.
When updating virtual columns when a row has been read, the allocated value
is stored in 'read_value' instead of 'value'.  The allocated blobs for the new
row is stored in 'value' as before.
I also made, as a safety precaution, the insert delayed handling of blobs
more general by using value to store strings instead of the record.
This ensures that virtual functions on delayed insert should work in as
in the case of normal insert.

The fix also ensures that we don't recalculate all virtual
columns for VCOL_UPDATE_FOR_READ_WRITE, only does that was
not updated by VCOL_UPDATE_FOR_READ.